Fix #718
