+++ stl/inc/condition_variable
@@ -124,12 +124,9 @@ public:
             return cv_status::timeout;
         }
 
-        // TRANSITION, ABI: The standard says that we should use a steady clock,
-        // but unfortunately our ABI relies on the system clock.
-        _timespec64 _Tgt;
-        const bool _Clamped     = _To_timespec64_sys_10_day_clamped(_Tgt, _Rel_time);
-        const cv_status _Result = _Wait_until_sys_time(_Lck, &_Tgt);
-        if (_Clamped) {
+        const auto _Rel_time_ms = _Clamped_rel_time_ms_count(_Rel_time);
+        const cv_status _Result = _Wait_for_ms_count(_Lck, _Rel_time_ms._Count);
+        if (_Rel_time_ms._Clamped) {
             return cv_status::no_timeout;
         }
 
@@ -206,12 +203,8 @@ public:
                 break;
             }
 
-            const auto _Rel_time = _Abs_time - _Now;
-            // TRANSITION, ABI: The standard says that we should use a steady clock,
-            // but unfortunately our ABI relies on the system clock.
-            _timespec64 _Tgt;
-            (void) _To_timespec64_sys_10_day_clamped(_Tgt, _Rel_time);
-            (void) _Cnd_timedwait(_Mycnd(), _Myptr->_Mymtx(), &_Tgt);
+            const unsigned long _Rel_ms_count = _Clamped_rel_time_ms_count(_Abs_time - _Now)._Count;
+            (void) _Cnd_timedwait_for(_Mycnd(), _Myptr->_Mymtx(), _Rel_ms_count);
             _Guard_unlocks_before_locking_outer.unlock();
         } // relock
 
@@ -234,12 +227,12 @@ private:
     }
 
     template <class _Lock>
-    cv_status _Wait_until_sys_time(_Lock& _Lck, const _timespec64* const _Abs_time) {
+    cv_status _Wait_for_ms_count(_Lock& _Lck, const unsigned _Rel_ms_count) {
         // wait for signal with timeout
         const shared_ptr<mutex> _Ptr = _Myptr; // for immunity to *this destruction
         unique_lock<mutex> _Guard{*_Ptr};
         _Unlock_guard<_Lock> _Unlock_outer{_Lck};
-        const _Thrd_result _Res = _Cnd_timedwait(_Mycnd(), _Ptr->_Mymtx(), _Abs_time);
+        const _Thrd_result _Res = _Cnd_timedwait_for(_Mycnd(), _Ptr->_Mymtx(), _Rel_ms_count);
         _Guard.unlock();
 
         if (_Res == _Thrd_result::_Success) {
