+++ stl/inc/header-units.json
@@ -17,6 +17,7 @@
         "__msvc_print.hpp",
         "__msvc_sanitizer_annotate_container.hpp",
         "__msvc_system_error_abi.hpp",
+        "__msvc_threads_core.hpp",        
         "__msvc_tzdb.hpp",
         "__msvc_xlocinfo_types.hpp",
         "algorithm",
